Add price API unittests
Remove the old tests and add some price API tests.
